Reorg composer, rework template editor
