+++ app/internal_packages/composer-templates/lib/template-actions.es6
@@ -1,0 +1,15 @@
+import Reflux from 'reflux';
+
+const TemplateActions = Reflux.createActions([
+  'insertTemplateId',
+  'createTemplate',
+  'showTemplates',
+  'deleteTemplate',
+  'renameTemplate',
+]);
+
+for (const key of Object.keys(TemplateActions)) {
+  TemplateActions[key].sync = true;
+}
+
+export default TemplateActions;
