+++ app/src/components/composer-editor/composer-support.es6
@@ -1,0 +1,8 @@
+module.exports = {
+  DraftJSConfig: require('./draftjs-config'),
+  InlineAttachmentPlugin: require('./inline-attachment-plugin'),
+  LinkifyPlugin: require('./linkify-plugin'),
+  QuotedTextPlugin: require('./quoted-text-plugin'),
+  TemplatesPlugin: require('./templates-plugin'),
+  TextStylePlugin: require('./text-style-plugin'),
+};
