+++ app/src/components/composer-editor/draftjs-config.jsx
