+++ app/src/components/composer-editor/templates-plugin.jsx
